Installation: Fix kernel module installation for custom DESTDIR
Issue:
When the user provides a custom path for the Makefile in $ROOT/kernel
directory, using the command "make install DESTDIR=<path>", it fails at
the stage when it is trying to copy the pf_ring.h header file over to
"<path>/usr/include/linux", if the afore-mentioned path doesn't exist.
The issue doesn't exist when it tries to copy the kernel module
pf_ring.ko over, since it has a "mkdir -p " for that, but the same
doesn't exist for copying the header file.

Solution:
Added a mkdir -p for the above "<path>/usr/include/linux.  There are no
errors if the path already exists.  If it doesn't exist, the path is
created.