Fix BPF inversion on Linux 4.4.x kernels
Linux sk_filter returns 0 if packet should be accepted and -EPERM
if the packet should be dropped. This leads into a situation where
current code causes all BPF filters to be inverse.

This commit fixes the inversion by setting res to 1 if sk_filter
yields an "accepted result", otherwise set res to 0.