when reader is slower than writer, buffer overwrite
Consider the following situation, A and B is slot_len exactly

1. write A

when
  insert_off = tot_mem - sizeof(FlowSlotInfo) - 1.9 * slot_len , and we simply represent insert_off as -1.9

so insert is permit,
  insert_off += slot_len, but insert_off no room for one another slot, then reset insert_off = 0.  write A in block [-1.9, -0.9)

2. read A

when
  remove_off -1.9,

pfring_recv() read A, then advance remove_off to 0, And block A [-1.9, -0.9) is used by userland program.

3. write B overlap with A

when userland reader is slower than writer, and -3 < insert_off < -2 , we assume insert_off as -2.1 simply.

Because, now remove_off is 0,  2 * slot_len < tot_mem - sizeof(FlowSlotInfo) - insert_off = 2.1 , So write B is permit.
  advance insert_off to -1.1, and write B in block [-2.1, -1.1).

Apparently B overwrite partial memory of A which is being processed by userland program.