kernel: Replace deprecated SUBDIRS with M in Makefile
The use of SUBDIRS to specify the location of external module source
files has been deprecated since Linux kernel version 5.3. The
recommended replacement is to use M instead.

This commit updates the Makefile to use M=$(HERE) instead of
SUBDIRS=$(HERE) in the 'clean' target. This change ensures compatibility
with newer Linux kernels and avoids potential issues during the build
process, such as the unintended deletion of files.